pyproject.toml
[tool.poetry]
name = "dcbot"
version = "0.1.0"
description = ""
authors = ["Kevin <kevintsai010111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
aiohttp = "^3.11.11"
python-dotenv = "^1.0.1"
pytz = "^2024.2"
requests = "^2.32.3"
yt-dlp = "^2025.1.15"
poetry-plugin-export = "^1.9.0"
pynacl = "^1.5.0"
brotli = "^1.1.0"
tqdm = "^4.67.1"
psutil = "^7.0.0"
flask = "^3.1.0"
python-socketio = "^5.13.0"
eventlet = "^0.39.1"
ffmpeg = "^1.4"
discord-py = {extras = ["voice"], version = "^2.5.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
